@@ -5,30 +5,31 @@
         "mongoose": "^6.0.0",
         "ws": "^5.0.0",
         "redis": "2.7.1",
-        "os-npm-util": "https://github.com/codeopensrc/os-npm-util/tarball/master#0.8.1"
-    },
-    "resolutions": {
-        "**/ua-parser-js": "0.7.30"
+        "os-npm-util": "https://github.com/codeopensrc/os-npm-util/tarball/master#0.10.0"
     },
     "devDependencies": {
-        "babel-core": "^6.22.1",
-        "babel-loader": "^6.2.4",
-        "babel-preset-es2015": "^6.22.0",
-        "babel-preset-react": "^6.22.0",
-        "babel-preset-stage-0": "^6.22.0",
-        "less": "^2.7.1",
-        "less-loader": "^2.2.3",
-        "css-loader": "^0.26.1",
-        "json-loader": "^0.5.4",
-        "react": "^15.3.0",
-        "react-dom": "^15.3.0",
-        "style-loader": "^0.13.1",
-        "webpack": "^1.14.0",
-        "html-webpack-plugin": "^2.30.1",
+        "@babel/core": "^7.15.5",
+        "@babel/preset-react": "^7.14.5",
+        "buffer": "^6.0.3",
+        "html-webpack-plugin": "^5.3.2",
+        "less": "^4.1.2",
+        "less-loader": "^10.0.1",
+        "babel-loader": "^8.2.1",
+        "css-loader": "^6.3.0",
+        "style-loader": "^3.3.0",
+        "exports-loader": "*",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
+        "react-router-dom": "^5.3.0",
+        "react-hot-loader": "^4.13.0",
+        "webpack": "^5.56.1",
+        "webpack-cli": "^4.8.0",
+        "webpack-dev-server": "^4.3.1",
         "marked": "^2.0.1"
     },
     "scripts": {
-        "watch": "webpack --config='src/config/webpack.config.js' --progress --colors --watch",
-        "release": "webpack --config='src/config/webpack.config.js' --progress --colors --optimize-minimize"
+        "reloader": "webpack serve --config='src/config/webpack.config.js' --color",
+        "watch": "webpack --config='src/config/webpack.config.js' --color --watch",
+        "release": "webpack --config='src/config/webpack.config.js' --color --optimization-minimize"
     }
 }
