@@ -1,46 +1,34 @@
 "use strict";
 
-const React = require('react');
-const DOM = require('react-dom');
-const { api } = require("os-npm-util")
+import React, { useEffect, useState, useCallback } from 'react';
+import { api } from "os-npm-util/client";
 
-const unity = require("../js/unity.js")
+import unity from "../js/unity.js";
 
-const Header = require('./Header.jsx');
+import Header from "./Header.jsx"
 
-require("../style/Games.less")
+import "../style/Games.less"
 
-const Unity = React.createClass({
+const Unity = function(props) {
+    const { baseAssetFolder } = props
+    const [title, setTitle] = useState("")
+    const [gamelist, setGamelist] = useState([])
+    const [gameactive, setGameactive] = useState(false)
 
-    getInitialState: function() {
-        return {
-            baseAssetFolder: this.props.baseAssetFolder,
-            title: "",
-            gamelist: [],
-            gameactive: false
-        };
-    },
-
-    componentWillUnmount: function() {
-        // Do cleanup
-    },
-
-    componentWillMount: function() {
+    useEffect(() => {
         api.get(`gamelist`, (res) => {
             if(res.status) {
-                this.setState({ gamelist: res.body })
+                setGamelist(res.body)
             }
             else {
                 console.log(res);
             }
         })
-    },
+        return api.abortAllRequests
+    }, [])
 
-    componentDidMount: function() {
-    },
-
-    loadGame: function(foldername, projectname) {
-        let fullprojectpath = `${this.state.baseAssetFolder}/${foldername}`
+    const loadGame = (foldername, projectname) => {
+        let fullprojectpath = `${baseAssetFolder}/${foldername}`
         let canvasContainer = document.querySelector("#unity-container");
 
         let oldcanvas = document.querySelector("#unity-canvas")
@@ -62,56 +50,53 @@ const Unity = React.createClass({
         document.head.appendChild(link1)
         document.head.appendChild(link2)
 
-        this.setState({
-            title: projectname,
-            gameactive: true
-        })
+        setTitle(projectname)
+        setGameactive(true)
         unity.load(foldername, fullprojectpath, projectname)
-    },
+    }
 
-    render: function() {
+    let gamebuttons = gamelist.map((game, i) => {
+        return (<button key={i} className={"games-btn"}
+                        onClick={() => loadGame(game.folder, game.projectname)}
+                        disabled={gameactive}>
+                    {game.projectname}
+                </button>)
+    })
 
-        let gamebuttons = this.state.gamelist.map((game, i) => {
-            return (<button key={i} className={"games-btn"}
-                            onClick={() => this.loadGame(game.folder, game.projectname)}
-                            disabled={this.state.gameactive}>
-                        {game.projectname}
-                    </button>)
-        })
-
-        return (
-            <div id={"gamespage"}>
-                <Header theme={"dark"}/>
-                <div id={"games-container"}>
-                    <div id={"games-sidebar"}>
-                        <span>Sample Selection</span>
-                        {gamebuttons}
-                        <button id={"unity-quit"}
-                                className={"games-btn"}
-                                onClick={() => this.setState({gameactive: false, title: ""})}
-                                disabled={!this.state.gameactive}>
-                            Quit
-                        </button>
+    return (
+        <div id={"gamespage"}>
+            <Header theme={"dark"}/>
+            <div id={"games-container"}>
+                <div id={"games-sidebar"}>
+                    <span>Sample Selection</span>
+                    {gamebuttons}
+                    <button id={"unity-quit"}
+                            className={"games-btn"}
+                            onClick={() => {
+                                setTitle("")
+                                setGameactive(false)
+                            }}
+                            disabled={!gameactive}>
+                        Quit
+                    </button>
+                </div>
+                <div id={"unity-container"} className={"unity-desktop"}>
+                    <canvas id={"unity-canvas"}></canvas>
+                    <div id={"unity-loading-bar"}>
+                        <div id={"unity-logo"}></div>
+                        <div id={"unity-progress-bar-empty"}>
+                            <div id={"unity-progress-bar-full"}></div>
+                        </div>
                     </div>
-                    <div id={"unity-container"} className={"unity-desktop"}>
-                        <canvas id={"unity-canvas"}></canvas>
-                        <div id={"unity-loading-bar"}>
-                            <div id={"unity-logo"}></div>
-                            <div id={"unity-progress-bar-empty"}>
-                                <div id={"unity-progress-bar-full"}></div>
-                            </div>
-                        </div>
-                        <div id={"unity-footer"}>
-                            <div id={"unity-webgl-logo"}></div>
-                            <div id={"unity-fullscreen-button"}></div>
-                            <div id={"unity-build-title"}>{this.state.title}</div>
-                        </div>
+                    <div id={"unity-footer"}>
+                        <div id={"unity-webgl-logo"}></div>
+                        <div id={"unity-fullscreen-button"}></div>
+                        <div id={"unity-build-title"}>{title}</div>
                     </div>
                 </div>
             </div>
-        )
-    }
+        </div>
+    )
+};
 
-});
-
-module.exports = Unity;
+export { Unity as default };
