@@ -1,43 +1,30 @@
 "use strict";
 
-import React from 'react';
-import DOM from 'react-dom';
+import React, { useEffect, useState, useCallback } from 'react';
 import marked from 'marked';
-import { api } from "os-npm-util";
+import { api } from "os-npm-util/client";
 
-const Header = require('./Header.jsx');
+import Header from "./Header.jsx"
 
-require("../style/Markdown.less")
+import "../style/Markdown.less"
 
-class Markdown extends React.Component {
+const Markdown = function(props) {
+    //const { } = props
+    const [markdown, setMarkdown] = useState("")
+    const [mdfiles, setMdfiles] = useState([])
 
-    static defaultProps = { };
-
-    constructor(props) {
-        super(props)
-        this.state = {
-            markdown: "",
-            mdfiles: []
-        }
-        this.renderMarkdownDir = this.renderMarkdownDir.bind(this);
-        this.renderMarkdown = this.renderMarkdown.bind(this);
-        this.getMarkdown = this.getMarkdown.bind(this);
-        this.buffToString = this.buffToString.bind(this);
-    }
-
-    componentDidMount() {
+    useEffect(() => {
         let url = location.pathname
-        this.renderMarkdownDir();
-        this.renderMarkdown(url)
-    }
+        //renderMarkdownDir();
+        renderMarkdown(url)
+        return api.abortAllRequests
+    }, [])
 
-    componentWillReceiveProps(nextProps) {
-    }
 
-    renderMarkdownDir() {
+    const renderMarkdownDir = () => {
         api.get(`/mddir`, (res) => {
             if(res.status) {
-                this.setState({mdfiles: res.body});
+                setMdfiles(res.body)
             }
             else {
                 console.log("Err fetching readmeDir");
@@ -45,24 +32,21 @@ class Markdown extends React.Component {
         })
     }
 
-    renderMarkdown(url) {
+    const renderMarkdown = (url) => {
         url === "/" ? url = "/Home" : ""
         url = `/md${url}`.replace(".md", "");
-        this.getMarkdown(url, (data) => {
-            this.setState({
-                markdown: data,
-            }, () => {
-                let formattedHash = location.hash.replace(/%20/g, "-").replace("#", "").toLowerCase();
-                formattedHash ? document.getElementById(formattedHash).scrollIntoView() : ""
-            })
+        getMarkdown(url, (data) => {
+            setMarkdown(data)
+            let formattedHash = location.hash.replace(/%20/g, "-").replace("#", "").toLowerCase();
+            formattedHash ? document.getElementById(formattedHash).scrollIntoView() : ""
         })
     }
 
-    getMarkdown(url, callback) {
+    const getMarkdown = (url, callback) => {
         api.get(url, (res) => {
             if(res.status) {
                 let buff = new Buffer(res.data.data)
-                let str = this.buffToString(buff)
+                let str = buffToString(buff)
                 let markedDownData = marked.parse(str)
                 callback(markedDownData)
             }
@@ -73,7 +57,7 @@ class Markdown extends React.Component {
     }
 
     // Handles large strings - More of a reference here and TextDecoder doesnt work in edge apparent
-    buffToString(buffer) {
+    const buffToString = (buffer) => {
         var bufView = new Uint16Array(buffer);
         var length = bufView.length;
         var result = '';
@@ -87,46 +71,41 @@ class Markdown extends React.Component {
         return result;
     }
 
-    render() {
-
-        let mdfiles = this.state.mdfiles.map((file, i) => {
-            return (
-                <div key={i} className={"filename"}>
-                    <a href={`/${file}`}>{file}</a>
-                </div>
-            )
-        })
-
-        let TOC = this.state.markdown ? this.state.markdown.matchAll(/h1 id="(.+)">(.+)</g) : [];
-        let links = Array.from(TOC, (header, i) => {
-            return (
-                <div key={i} className={"header-toc"}>
-                    <a href={`#${header[1]}`}>{header[2]}</a>
-                </div>
-            )
-        })
-        //<div id="filenav" className={"sidebars"}>
-        //    <span>Topics:</span>
-        //    {mdfiles}
-        //</div>
-
+    let renderMdfiles = mdfiles.map((file, i) => {
         return (
-            <div id="component-markdown">
-                <Header />
-
-                <div id="markdown-container">
-                    <div id="markdown" dangerouslySetInnerHTML={{__html: this.state.markdown}} />
-                    <div id="toc" className={"sidebars"}>
-                        <span>Table of Contents</span>
-                        {links}
-                    </div>
-                </div>
-
+            <div key={i} className={"filename"}>
+                <a href={`/${file}`}>{file}</a>
             </div>
-        );
-    }
+        )
+    })
 
+    let TOC = markdown ? markdown.matchAll(/h1 id="(.+)">(.+)</g) : [];
+    let links = Array.from(TOC, (header, i) => {
+        return (
+            <div key={i} className={"header-toc"}>
+                <a href={`#${header[1]}`}>{header[2]}</a>
+            </div>
+        )
+    })
+
+    //<div id="filenav" className={"sidebars"}>
+    //    <span>Topics:</span>
+    //    {renderMdfiles}
+    //</div>
+
+    return (
+        <div id="component-markdown">
+            <Header />
+
+            <div id="markdown-container">
+                <div id="markdown" dangerouslySetInnerHTML={{__html: markdown}} />
+                <div id="toc" className={"sidebars"}>
+                    <span>Table of Contents</span>
+                    {links}
+                </div>
+            </div>
+        </div>
+    );
 }
 
-module.exports = Markdown
-
+export { Markdown as default };
